requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Dropzone==1.6.0
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.0.3
Flask-Uploads==0.2.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.3
six==1.16.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
Werkzeug==2.3.4
